Trait ssdp::header::HeaderRef

pub trait HeaderRef: Debug {
    fn get<H>(&self) -> Option<&H>
    where
        H: Header + HeaderFormat;

    fn get_raw(&self, name: &str) -> Option<&[Vec<u8>]>;
}

Trait for viewing the contents of a header structure.

Required Methods

fn get<H>(&self) -> Option<&H> where
H: Header + HeaderFormat,

View a reference to a header field if it exists.

fn get_raw(&self, name: &str) -> Option<&[Vec<u8>]>

View a reference to the raw bytes of a header field if it exists.