Module ssdp::message

Messaging primitives for discovering devices and services.

Structs

NotifyListener
Notify listener that can listen to notify messages sent within the network.

NotifyMessage
Notify message that can be sent via multicast to devices on the network.

SearchListener
Search listener that can listen for search messages sent within the network.

SearchRequest
Search request that can be sent via unicast or multicast to devices on the network.

SearchResponse
Search response that can be received or sent via unicast to devices on the network.

Enums

MessageType
Enumerates different types of SSDP messages.

Constants

UPNP_MULTICAST_PORT

Traits

Listen