Struct ssdp::SSDPReceiver

pub struct SSDPReceiver<T> {
    // some fields omitted
}

A non-blocking SSDP message receiver.

Methods

impl<T> SSDPReceiver<T> where T: FromRawSSDP + Send + 'static

fn new(socks: Vec<UdpSocket>, time: Option<Duration>) -> Result<SSDPReceiver<T>>

Construct a receiver that receives bytes from a number of UdpSockets and tries to construct an object T from them. If a duration is provided, the channel will be shutdown after the specified duration.

Due to implementation details, none of the UdpSockets should be bound to the default route, 0.0.0.0, address.

impl<T> SSDPReceiver<T>

fn try_recv(&self) -> Result<(T, SocketAddr), TryRecvError>

Non-blocking method that attempts to read a value from the receiver.

fn recv(&self) -> Result<(T, SocketAddr), RecvError>

Blocking method that reads a value from the receiver until one is available.