Crate ssdp

An asynchronous abstraction for discovering devices and services on a network.

SSDP stands for Simple Service Discovery Protocol and it is a protocol that uses HTTPMU to distribute messages across a local network for devices and services to discover each other. SSDP can most commonly be found in devices that implement UPnP as it is used as the discovery mechanism for that standard.

Modules

header
Headers and primitives for parsing headers within SSDP requests.

message
Messaging primitives for discovering devices and services.

Structs

SSDPError
The Error type.

SSDPIter
Iterator for an SSDPReceiver.

SSDPReceiver
A non-blocking SSDP message receiver.

Enums

FieldMap
Enumerates key value pairs embedded within SSDP header fields.

SSDPErrorKind
The kind of an error.

Traits

SSDPResultExt
Additionnal methods for Result, for easy interaction with this crate.

Type Definitions

SSDPResult
Convenient wrapper around std::Result.