Trait ssdp::message::listen::Listen

pub trait Listen {
    type Message: FromRawSSDP + Send + 'static;
    fn listen() -> SSDPResult<SSDPReceiver<Self::Message>> { ... }

    fn listen_with_config(
        config: &Config
    ) -> SSDPResult<SSDPReceiver<Self::Message>> { ... }
}

Associated Types

type Message: FromRawSSDP + Send + 'static

Provided Methods

fn listen() -> SSDPResult<SSDPReceiver<Self::Message>>

Listen for messages on all local network interfaces.

This will call listen_with_config() with default values.

fn listen_with_config(
config: &Config
) -> SSDPResult<SSDPReceiver<Self::Message>>

Listen for messages on all local network interfaces.

Notes

This will bind to each interface, NOT to INADDR_ANY.

If you are on an environment where the network interface will be changing, you will have to stop listening and start listening again, or we recommend using listen_anyaddr_with_config() instead.