Struct ssdp::message::SearchRequest

pub struct SearchRequest { /* fields omitted */ }

Search request that can be sent via unicast or multicast to devices on the network.

Methods

impl SearchRequest

fn new() -> SearchRequest

Construct a new SearchRequest.

fn unicast<A: ToSocketAddrs>(&mut self,
dst_addr: A)
-> SSDPResult<SSDPReceiver<SearchResponse>>

Send this search request to a single host.

Currently this sends the unicast message on all available network interfaces. This assumes that the network interfaces are operating on either different subnets or different ip address ranges.

fn multicast(&mut self) -> SSDPResult<SSDPReceiver<SearchResponse>>

Send this search request to the standard multicast address:port.

fn multicast_with_port(&mut self,
port: u16)
-> SSDPResult<SSDPReceiver<SearchResponse>>

Send this search request to the standard multicast address but a custom port